Add a dev server for the website

+++ web/build.sh
@@ -18,20 +18,27 @@
 
 export LC_ALL=en_US.UTF-8
 export LANG=en_US.UTF-8
-root=_site
 
-cd data; git pull; cd ..
+if [ $1 == "--quick" ]
+then root="."
+else root="_site"
+fi
 
-rm -rf $root
-mkdir $root
-mkdir $root/translations
-cp -r assets    $root/assets
-cp -r data      $root/data
+# cd data; git pull; cd ..
 
-copy_files=(index.html body.html)
-for file in ${copy_files[*]}; do
-    cp $file $root/$file
-done
+if [ $1 != "--quick" ]
+then
+    rm -rf $root
+    mkdir $root
+    mkdir $root/translations
+    cp -r assets    $root/assets
+    cp -r data      $root/data
+
+    copy_files=(index.html body.html)
+    for file in ${copy_files[*]}; do
+        cp $file $root/$file
+    done
+fi
 
 sass css/index.sass $root/index.css
 cat js/lib/*.js         >> $root/main.js
